小改

+++ types.h
@@ -1,3 +1,4 @@
 #include <vector>
 #include <map>
-typedef std::map<int, std::vector<std::pair<int, float> > > DataSet;
+typedef std::map<int, std::vector<std::pair<int, double> > > DataSet;
+/* typedef std::vector<std::pair<int,std::vector<std::pair<int,float> > > > DataSet; */
